+++ src/CronScheduler.Extensions/Scheduler/IScheduledJob.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Threading;
+﻿using System.Threading;
 using System.Threading.Tasks;
 
 namespace CronScheduler.Extensions.Scheduler
@@ -10,22 +9,11 @@ namespace CronScheduler.Extensions.Scheduler
     public interface IScheduledJob
     {
         /// <summary>
-        /// Cron schedule.
+        /// The name of the executing job.
+        /// In order for the <see cref="SchedulerOptions"/> options to work correctly make sure that the name is matched
+        /// between the job and the named job options.
         /// </summary>
-        [Obsolete("This property will be remove from the interface in the next release. Please use" + nameof(SchedulerOptions.CronSchedule) + " instead.")]
-        string CronSchedule { get; }
-
-        /// <summary>
-        /// When not specified the CronScheduler will use Local Time.
-        /// </summary>
-        [Obsolete("This property will be remove from the interface in the next release. Please use" + nameof(SchedulerOptions.CronTimeZone) + " instead.")]
-        string CronTimeZone { get; }
-
-        /// <summary>
-        /// Should be run on application start.
-        /// </summary>
-        [Obsolete("This property will be remove from the interface in the next release. Please use" + nameof(SchedulerOptions.RunImmediately) + " instead.")]
-        bool RunImmediately { get; }
+        string Name { get; }
 
         /// <summary>
         /// Job that will be executing on this schedule.