+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 Change Log
 ===============================================================================
 
+Version 3.0.0
+
+* Removed ScheduledJob class
+* Removed properties from IScheduledJob (CronSchedule, CronTimeZone, RunImmediately)
+* Simplify registration for the jobs and options.
+* Allow for the same job type to be registered with different cron schedule options.
+* Jobs options are refreshed if the configuration provider notifies of the changes.
+
+Version 2.1.0
+
+* Introduced the ability to register schedule jobs at any point of time not only at the startup.
+* Introduced the `SchedulerOptions` mechanism to update the options on configuration provider reload.
+
 Version 1.1.2
 
 * Resolved issue #22 "schedule executing twice per day instead of once"